Revert "MOL-426: CLI shipping for all orders"
This reverts commit f0324bf4250c8b83974a1d04b6e4cd807bf2c250.
