--- Command/ShippingCommand.php
@@ -1,112 +1,0 @@
-<?php
-
-namespace MollieShopware\Command;
-
-use Doctrine\ORM\EntityManager;
-use MollieShopware\Components\Config;
-use MollieShopware\Facades\FinishCheckout\Services\MollieStatusValidator;
-use MollieShopware\Facades\ShippingCommand\ShippingCommandFacade;
-use MollieShopware\Gateways\MollieGatewayInterface;
-use MollieShopware\Models\Transaction;
-use Psr\Log\LoggerInterface;
-use Shopware\Commands\ShopwareCommand;
-use Shopware\Components\Model\ModelManager;
-use Shopware\Models\Order\Order;
-use Shopware\Models\Shop\Shop;
-use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\Console\Style\SymfonyStyle;
-
-class ShippingCommand extends ShopwareCommand
-{
-
-    /**
-     *
-     */
-    const LOG_PREFIX = 'CLI Shipping: ';
-
-    /**
-     * @var ShippingCommandFacade
-     */
-    private $facade;
-
-    /**
-     * @var EntityManager
-     */
-    private $entityManager;
-
-
-    /**
-     * @param Config $config
-     * @param ModelManager $modelManager
-     * @param MollieGatewayInterface $gwMollie
-     * @param LoggerInterface $logger
-     * @param null $name
-     */
-    public function __construct(Config $config, ModelManager $modelManager, MollieGatewayInterface $gwMollie, LoggerInterface $logger, $name = null)
-    {
-        parent::__construct($name);
-
-        $this->entityManager = $modelManager;
-
-        $repoShops = $modelManager->getRepository(Shop::class);
-        $repoOrders = $modelManager->getRepository(Order::class);
-        $repoTransactions = $modelManager->getRepository(Transaction::class);
-
-        $this->facade = new ShippingCommandFacade(
-            self::LOG_PREFIX,
-            $config,
-            $gwMollie,
-            new MollieStatusValidator(),
-            $logger,
-            $repoShops,
-            $repoOrders,
-            $repoTransactions
-        );
-    }
-
-    /**
-     *
-     */
-    public function configure()
-    {
-        $this
-            ->setName('mollie:ship:orders')
-            ->setDescription('Ship completed orders');
-    }
-
-    /**
-     * @param InputInterface $input
-     * @param OutputInterface $output
-     * @return int|void|null
-     * @throws \Exception
-     */
-    public function execute(InputInterface $input, OutputInterface $output)
-    {
-        $io = new SymfonyStyle($input, $output);
-        $io->title('MOLLIE Ship Command');
-        $io->text('Searching for all non-shipped orders and mark them as shipped if the status is correct...');
-
-
-        $qb = $this->entityManager->createQueryBuilder();
-
-        /** @var Transaction[] $transactions */
-        $transactions = $qb->select('t')
-            ->from(Transaction::class, 't')
-            ->where($qb->expr()->like('t.mollieId', ':mollieId'))
-            ->andWhere($qb->expr()->eq('t.isShipped', ':shipped'))
-            ->setParameter(':mollieId', 'ord_%')
-            ->setParameter(':shipped', false)
-            ->getQuery()
-            ->getResult();
-
-
-        if ($transactions === null || !is_array($transactions)) {
-            $io->success("No Mollie Transactions found!");
-            return;
-        }
-
-        $this->facade->ship($transactions, $io, $output, $this->container);
-    }
-
-}
